ReservaMe-Backend/reservame-backend/requirements.txt
Flask
flask-blueprints
pyodbc
bcrypt
python-dotenv

#pip install -r requirements.txt
